requirements.txt
asgiref==3.2.10
Django==3.1.2
django-bootstrap-form==3.4
django-crispy-forms==1.9.2
django-filter==2.4.0
django-localflavor==3.0.1
django-multiselectfield==0.1.12
django-widget-tweaks==1.4.8
gunicorn==20.0.4
Pillow==7.2.0
python-stdnum==1.14
pytz==2020.1
register==0.1
six==1.15.0
sqlparse==0.3.1
whitenoise==5.2.0
